Workout list refactoring (change method, variable names), fix toast messages

+++ src/views/TheWorkoutView.vue
@@ -18,7 +18,7 @@
         <v-btn
           icon
           color="pink"
-          @click.stop="addItem"
+          @click.stop="validateAddInput"
         >
           <v-icon>mdi-plus</v-icon>
         </v-btn>
@@ -33,49 +33,57 @@
         class="overflow-y-auto"
       >
         <v-list>
-          <draggable
-            v-model="workouts"
-            group="people"
-            @end="onDragEnd"
+          <v-list-item
+            v-for="element in workouts"
+            :key="element.id"
+            style="cursor: pointer; text-align: left"
           >
-            <v-list-item
-              v-for="element in workouts"
-              :key="element.id"
-              style="cursor: pointer; text-align: left"
-            >
-              <v-list-item-title>{{ element.title }}</v-list-item-title>
-              <v-list-item-avatar>
-                <v-icon
-                  v-blur
-                  @click.stop="removeWorkout(element.id)"
-                  v-text="`mdi-minus-circle-outline`"
-                />
-              </v-list-item-avatar>
-            </v-list-item>
-          </draggable>
+            <v-list-item-title>{{ element.title }}</v-list-item-title>
+            <v-list-item-avatar>
+              <v-icon
+                v-blur
+                @click.stop="openRemoveWorkoutModal(element.id)"
+                v-text="`mdi-minus-circle-outline`"
+              />
+            </v-list-item-avatar>
+          </v-list-item>
         </v-list>
       </v-col>
     </v-row>
     <the-confirm-dialog
+      v-if="currentDialog !== ''"
       v-model="dialog"
-      :title="`Delete workout`"
-      :description="`Are you sure you want to delete?`"
+      :title="dialogs[currentDialog].title"
+      :description="dialogs[currentDialog].description"
       @close="closeDialog"
     />
   </v-container>
 </template>
 <script>
-import draggable from 'vuedraggable'
-import remove from 'lodash/remove'
 import TheConfirmDialog from "@/components/dialogs/TheConfirmDialog";
+import {find} from "lodash";
 
+const DIALOGS = {
+  DELETE: 'delete',
+  ADD: 'add'
+}
 export default {
   components: {
     TheConfirmDialog,
-    draggable
   },
   data() {
     return {
+      dialogs: {
+        delete: {
+          title: 'Delete workout',
+          description: 'Are you sure you want to delete?'
+        },
+        add: {
+          title: 'Add workout',
+          description: 'Do you want to add this workout?'
+        }
+      },
+      currentDialog: '',
       selectedWorkout: '',
       dialog: false,
       newWorkout: '',
@@ -87,34 +95,48 @@ export default {
     this.workouts = await this.getWorkouts()
   },
   methods: {
+    closeDialog(confirm) {
+      this.dialog = false
+      if (this.currentDialog === DIALOGS.ADD) {
+          if (confirm) {
+            this.addWorkout({title: this.newWorkout})
+            this.newWorkout = ''
+          }
+      } else if (this.currentDialog === DIALOGS.DELETE) {
+        if (confirm) {
+          this.deleteWorkout()
+        } else {
+          this.selectedWorkout = ''
+        }
+      }
+      this.currentDialog = ''
+    },
+    openRemoveWorkoutModal(id) {
+      this.currentDialog = DIALOGS.DELETE
+      this.selectedWorkout = id
+      this.dialog = true
+    },
+    validateAddInput() {
+      if (this.newWorkout === '') {
+        return
+      }
+      if (find(this.workouts, {title: this.newWorkout})) {
+        this.$toast.error('Workout already exists')
+        return
+      }
+      this.openAddConfirmModal()
+    },
+    openAddConfirmModal() {
+      this.currentDialog = DIALOGS.ADD
+      this.dialog = true
+    },
     async getWorkouts() {
       const {data} = await this.$api.workout.workouts()
       return data
     },
-    closeDialog(confirm) {
-      if (confirm) {
-        this.dialog = false
-        this.deleteWorkout()
-      } else {
-        this.selectedWorkout = ''
-        this.dialog = false
-      }
-    },
-    onDragEnd(v) {
-      if (v.oldIndex !== v.newIndex) {
-        console.info(`position moved from ${v.oldIndex} to ${v.newIndex}`)
-        // todo: change order API
-      }
-    },
-    removeWorkout(id) {
-      // todo: remove item with id
-      console.info('remove workout --', id)
-      this.dialog = true
-      this.selectedWorkout = id
-    },
     async deleteWorkout() {
       const {data} = await this.$api.workout.deleteWorkout(this.selectedWorkout)
-      if (data.showYn === 'N') {
+      if (data.show_yn === 'N') {
         this.$toast(`Workout removed`);
         this.workouts = await this.getWorkouts()
       } else {
@@ -122,25 +144,17 @@ export default {
       }
 
     },
-    async addItem() {
-      if (this.newWorkout === '') {
-        return
-      }
-
-      const r = await this.addWorkout({title: this.newWorkout})
-      if (r.title !== '') {
-        this.workouts = await this.getWorkouts()
-        this.$toast('Workout created')
-      } else {
-        this.$toast.error('failed to create workout')
-      }
-      this.newWorkout = ''
-    },
     async addWorkout(params) {
       try {
         const {data} = await this.$api.workout.AddWorkout(params)
-        return data
+        if (data.title !== '') {
+          this.workouts = await this.getWorkouts()
+          this.$toast('Workout created')
+        } else {
+          this.$toast.error('failed to create workout')
+        }
       } catch (e) {
+        this.$toast.error('API ERROR')
         return e
       }
     }
